% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_im}
\alias{get_im}
\title{Get images}
\source{
\url{https://a.mapillary.com/#images}
}
\usage{
get_im(key, size = "m", save)
}
\arguments{
\item{key}{Image key.}

\item{size}{Image size. One of \code{s[mall]} (320px), 
\code{m[edium]} (640px, the default), \code{l[arge]} (1024px) 
or \code{h[uge]} (2048px).}

\item{save}{Directory where to save the image file.}
}
\value{
An image.
}
\description{
Get and view images.
}
\examples{
\dontrun{
search_im(key=, size="m")
}
}

