% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_img}
\alias{get_img}
\title{Get images}
\source{
\url{https://a.mapillary.com/#images}
}
\usage{
get_img(img_key, size = "m", dir = tempdir(), display = TRUE)
}
\arguments{
\item{img_key}{Image key or vector of image keys.}

\item{size}{Image size. One of \code{s[mall]} (320px), 
\code{m[edium]} (640px, the default), \code{l[arge]} (1024px) 
or \code{h[uge]} (2048px).}

\item{dir}{Directory where to save the image file (default is a
temporary directory).}

\item{display}{If \code{TRUE} (default), the image is displayed 
If \code{img_key} is a vector, only the first image is displayed.}
}
\value{
The path to the saved image(s) as string or vector of strings.
}
\description{
Save and display images in R.
}
\examples{
\dontrun{
img <- images(closeto=c(9.436385,46.336591), radius=1000, 
  page=1, per_page=1, print=FALSE)$img_key
get_img(img_key=img)
img.path <- get_img(img_key=img, size="h")
img.path
}
}
