% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{images}
\alias{images}
\title{Search for images}
\source{
\url{https://a.mapillary.com/#images}
}
\usage{
images(bbox, closeto, radius, lookat, start_time, end_time, user_name, user_key,
  project_key, page, per_page, fields, print = TRUE)
}
\arguments{
\item{bbox}{Bounding box, given as vector of minx, miny, maxx, maxy.}

\item{closeto}{Location that images are close to, given as vector of longitude and latitude.}

\item{radius}{Radius around the \code{closeto} location, given in meters (default 100).}

\item{lookat}{Location in which direction images are taken (and therefore that location 
is likely to be visible in the images), given as vector of longitude and latitude.}

\item{start_time}{Start time images are captured (following ISO 8601 rules).}

\item{end_time}{End time images are captured before (following ISO 8601 rules).}

\item{user_name}{Just objects for specific users, given as vector of usernames.}

\item{user_key}{Just objects for specific users, given as vector of user keys.}

\item{project_key}{Just objects for specific projects, given as vector of project keys.}

\item{page}{Page number in pagination.}

\item{per_page}{Results per page in pagination.}

\item{fields}{Partially selected output fields, given as string or vector of strings. 
Fields are sorted in given order. Available fields: \code{camera_angle}, 
\code{camera_make}, \code{camera_model}, \code{captured_at}, \code{img_key}, 
\code{panorama}, \code{user_key}, \code{user_name}, \code{project_key}, \code{longitude}, \code{latitude}. 
If \code{fields} is missing (default), all available fields are returned.}

\item{print}{If \code{TRUE} (default) the search results are printed.}
}
\value{
A \code{data.frame} of matching images.
}
\description{
Search for images.
}
\examples{
\dontrun{
images(bbox=c(19.963211,49.317328,20.004066,49.325832), page=1, per_page=10)
}
}
