% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapillRy.R
\docType{package}
\name{mapillRy}
\alias{mapillRy}
\alias{mapillRy-package}
\alias{mapillRy-package}
\title{R Interface to the Mapillary API}
\description{
A wrapper for the Mapillary API. 'mapillRy' gives access to images, 
image sequences and stats of Mapillary, a service for crowdsourcing map images: 
<https://www.mapillary.com>.

Mapillary images are available under the Creative Commons Attribution-ShareAlike 4.0 
International License (CC BY-SA 4.0).
}
