% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gets.R
\name{search_im_cm}
\alias{search_im_cm}
\title{Search for commented images}
\source{
\url{https://a.mapillary.com/#get-searchimcm}
}
\usage{
search_im_cm(min_lat, max_lat, min_lon, max_lon, start_time, end_time, user,
  limit, page, print = TRUE)
}
\arguments{
\item{min_lat}{Minimum latitude.}

\item{max_lat}{Maximum latitude.}

\item{min_lon}{Minimum longitude.}

\item{max_lon}{Maximum longitude.}

\item{start_time}{Start time in EPOCH ms.}

\item{end_time}{End time in EPOCH ms.}

\item{user}{Just objects for specific user.}

\item{limit}{Results per page in pagination.}

\item{page}{Page number in pagination.}

\item{print}{if \code{TRUE} (default) the search results are printed.}
}
\value{
A \code{data.frame} of matching images.
}
\description{
Search for images that have at least one comment in an area.
}
\examples{
\dontrun{
search_im_cm(min_lat=49.317328, max_lat=49.325832,
  min_lon=19.963211, max_lon=20.004066)
}
}

