% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{search_im_random}
\alias{search_im_random}
\title{Get random image}
\source{
\url{https://a.mapillary.com/#get-searchimrandomselected}
}
\usage{
search_im_random(print = TRUE)
}
\arguments{
\item{print}{if \code{TRUE} (default) the search results are printed.}
}
\value{
A \code{data.frame} of the random image.
}
\description{
Get a random image from the Mapillary image archive, 
the images is randomized from a set of curated images.
}
\examples{
\dontrun{
search_im_random()
}
}

