% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gets.R
\name{search_seq}
\alias{search_seq}
\title{Search for sequences}
\source{
\url{https://a.mapillary.com/#get-searchsul}
}
\usage{
search_seq(min_lat, max_lat, min_lon, max_lon, start_time, end_time, user,
  limit, page, print = TRUE)
}
\arguments{
\item{min_lat}{Minimum latitude.}

\item{max_lat}{Maximum latitude.}

\item{min_lon}{Minimum longitude.}

\item{max_lon}{Maximum longitude.}

\item{start_time}{Start time in EPOCH ms.}

\item{end_time}{End time in EPOCH ms.}

\item{user}{Just objects for specific user.}

\item{limit}{Results per page in pagination.}

\item{page}{Page number in pagination.}

\item{print}{if \code{TRUE} (default) the search results are printed.}
}
\value{
A \code{data.frame} of matching sequences.
}
\description{
Search for sequences.
}
\examples{
\dontrun{
search_seq(min_lat=49.019063, max_lat=49.328211,
  min_lon=-123.233919, max_lon=-122.298288)
}
}

