% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{search_user}
\alias{search_user}
\title{Search for usernames}
\source{
\url{https://a.mapillary.com/#get-searchu}
}
\usage{
search_user(user, print = TRUE)
}
\arguments{
\item{user}{String to try to match for a username.}

\item{print}{if \code{TRUE} (default) the search results are printed.}
}
\value{
A \code{data.frame} of the matching usernames.
}
\description{
Search for usernames.
}
\examples{
\dontrun{
search_user("mynameis")
}
}
