% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{sequences}
\alias{sequences}
\title{Search for sequences}
\source{
\url{https://a.mapillary.com/#sequences}
}
\usage{
sequences(bbox, start_time, end_time, user_name, user_key, starred = FALSE,
  page, per_page, fields, print = TRUE)
}
\arguments{
\item{bbox}{Bounding box, given as vector of minx, miny, maxx, maxy.}

\item{start_time}{Start time images are captured (following ISO 8601 rules).}

\item{end_time}{End time images are captured before (following ISO 8601 rules).}

\item{user_name}{Just objects for specific users, given as vector of usernames.}

\item{user_key}{Just objects for specific users, given as vector of user keys.}

\item{starred}{If \code{TRUE}, only starred sequences are requested. Default is \code{FALSE}.}

\item{page}{Page number in pagination.}

\item{per_page}{Results per page in pagination.}

\item{fields}{Partially selected output fields, given as string or vector of strings. 
Available fields: \code{camera_angle}, \code{camera_make}, \code{camera_model}, 
\code{captured_at}, \code{img_key}, \code{panorama}, \code{user_key}, \code{user_name}, 
\code{longitude}, \code{latitude}. Default is all available fields.}

\item{print}{if \code{TRUE} (default) the search results are printed.}
}
\value{
A \code{data.frame} of matching sequences.
}
\description{
Search for sequences.
}
\examples{
\dontrun{
sequences(bbox=c(19.963211,49.317328,20.004066,49.325832), page=1, per_page=10)
}
}
