% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{stats_im}
\alias{stats_im}
\title{Statistics about images in the Mapillary system}
\source{
\url{https://a.mapillary.com/#get-statsim}
}
\usage{
stats_im(print = TRUE)
}
\arguments{
\item{print}{if \code{TRUE} (default) the statistics are printed.}
}
\value{
A \code{data.frame} with basic statistics.
}
\description{
Retrieve statistics about images in the Mapillary system.
}
\examples{
\dontrun{
stats_im()
}
}

