% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{stats_top}
\alias{stats_top}
\title{Top lists statistics for Mapillary}
\source{
\url{https://a.mapillary.com/#get-statsimtoplist}
}
\usage{
stats_top(cname, limit, print = TRUE)
}
\arguments{
\item{cname}{Name of country for stats (optional - if not added whole world is assumed).}

\item{limit}{Number of users in every list (default 10, optional).}

\item{print}{if \code{TRUE} (default) the statistics are printed.}
}
\value{
A \code{list} of two \code{data.frame}s, one for each toplist.
}
\description{
Retrieve weekly and total top lists for Mapillary.
}
\examples{
\dontrun{
stats_top("namibia")
}
}

