% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{user_stats}
\alias{user_stats}
\title{Get user stats}
\source{
\url{https://a.mapillary.com/#the-user-statistics}
}
\usage{
user_stats(user_name, user_key, fields, json = FALSE, print = TRUE)
}
\arguments{
\item{user_name}{Username as string (or vector of strings). Optional, if \code{user_key} is given.}

\item{user_key}{User key as string (or vector of strings). Optional, if \code{user_name} is given.}

\item{fields}{Partially selected output fields, given as string or vector of strings. 
Fields are sorted in given order. Available fields: \code{user_name}, \code{user_key},
\code{images}, \code{sequences}, \code{edits}, \code{blurs}. 
If \code{fields} is missing (default), all available fields are returned.}

\item{json}{If \code{FALSE} (default) the results are returned as simplified
\code{data.frame}. \code{TRUE} (invisibly) returns the original JSON object (\code{fields} is
ignored.}

\item{print}{if \code{TRUE} (default) the search results are printed.}
}
\value{
A \code{data.frame} of user statistics.
}
\description{
Get statistics about a user.
}
\examples{
\dontrun{
user_stats(user_name="billy_bob")
user_stats(user_name=c("billy_bob", "mthagaard"), 
           user_key=c("5gXh9Bb43yNhWOCoVC-FjQ", "vRyJQKolUExxn6HQiTZMRg"), 
           fields=c("user_name","blurs","images", "distance"))
}
}
