% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{users}
\alias{users}
\title{Search for users}
\source{
\url{https://a.mapillary.com/#users}
}
\usage{
users(bbox, user_name, user_key, page, per_page, fields, json = FALSE,
  print = TRUE)
}
\arguments{
\item{bbox}{Bounding box, given as vector of minx, miny, maxx, maxy.}

\item{user_name}{String (or vector of strings) to filter for usernames.}

\item{user_key}{String (or vector of strings) to filter for user keys.}

\item{page}{Page number in pagination.}

\item{per_page}{Results per page in pagination.}

\item{fields}{Partially selected output fields, given as string or vector of strings. 
Fields are sorted in given order. Available fields: \code{about}, 
\code{avatar}, \code{created_at}, \code{key}, \code{username}. 
If \code{fields} is missing (default), all available fields are returned.}

\item{json}{If \code{FALSE} (default) the results are returned as simplified
\code{data.frame}. \code{TRUE} (invisibly) returns the original JSON object (\code{fields} is
ignored.}

\item{print}{if \code{TRUE} (default) the search results are printed.}
}
\value{
A \code{data.frame} of users.
}
\description{
Search for users who contributed in a certain region.
}
\details{
Returned users are ordered by \code{created_at}. If \code{bbox} is provided, 
users are ordered by their last captured times.
}
\examples{
\dontrun{
users(user_name="billy_bob")
}
}
