% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{view_im}
\alias{view_im}
\title{View images}
\source{
\url{https://a.mapillary.com/#images}
}
\usage{
view_im(key, mode = "m", size = "m")
}
\arguments{
\item{key}{Image key.}

\item{mode}{Image view mode. Use \code{m[apillary]} (the default), 
to open the image in the mapillary website image view. 
Use \code{[d]irect}, to open the direct image link.}

\item{size}{Image size. Used only for \code{direct} image view mode.
One of \code{s[mall]} (320px), \code{m[edium]} (640px, the default), 
\code{l[arge]} (1024px) or \code{h[uge]} (2048px).}
}
\value{
An image.
}
\description{
View images online.
}
\examples{
\dontrun{
img <- search_im_close(lat=46.804159, lon=7.166325, 
  distance=10000, limit=1, print=FALSE)$key
view_im(key=img)
}
}

