% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{view_img}
\alias{view_img}
\title{View images}
\source{
\url{https://a.mapillary.com/#images}
}
\usage{
view_img(img_key, mode = "m", size = "m")
}
\arguments{
\item{img_key}{Image key.}

\item{mode}{Image view mode. Use \code{m[apillary]} (the default), 
to open the image in the mapillary website image view. 
Use \code{[d]irect}, to open the direct image link.}

\item{size}{Image size. Used only for \code{direct} image view mode.
One of \code{s[mall]} (320px), \code{m[edium]} (640px, the default), 
\code{l[arge]} (1024px) or \code{h[uge]} (2048px).}
}
\value{
A string containing the url to the image.
}
\description{
View images online.
}
\examples{
\dontrun{
img <- images(closeto=c(9.436385,46.336591), radius=1000, 
  page=1, per_page=1, print=FALSE)$img_key
view_img(img_key=img)
}
}
