% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_img}
\alias{get_img}
\title{Get images}
\source{
\url{https://a.mapillary.com/#images}
}
\usage{
get_img(img_key, size = "m", save)
}
\arguments{
\item{img_key}{Image key.}

\item{size}{Image size. One of \code{s[mall]} (320px), 
\code{m[edium]} (640px, the default), \code{l[arge]} (1024px) 
or \code{h[uge]} (2048px).}

\item{save}{Directory where to save the image file. Optional --
if missing, the image is saved as temporary file and diplayed.}
}
\value{
If \code{save=FALSE}, an image is displayed only. 
Otherwise, the path of the saved image is returned.
}
\description{
Save images and display images in R.
}
\examples{
\dontrun{
img <- images(closeto=c(9.436385,46.336591), radius=1000, 
  page=1, per_page=1, print=FALSE)$img_key
get_img(img_key=img)
img.path <- get_img(img_key=img, size="h", save=tempdir())
img.path
}
}
