% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapillRy.R
\docType{package}
\name{mapillRy}
\alias{mapillRy}
\alias{mapillRy-package}
\title{R Interface to the Mapillary API}
\description{
A wrapper for the Mapillary API,
available at \url{http://a.mapillary.com}. 'mapillRy' gives access to images,
image sequences and stats of Mapillary, a service for crowdsourcing map images.
A personal token is required for usage.
}

